Cleanup

+++ components/msg/constellation_msg.rs
@@ -259,8 +259,8 @@ impl LoadData {
 
 #[derive(Clone, PartialEq, Eq, Copy, Hash, Debug, Deserialize, Serialize)]
 pub enum NavigationDirection {
-    Forward(u32),
-    Back(u32),
+    Forward(usize),
+    Back(usize),
 }
 
 #[derive(Clone, PartialEq, Eq, Copy, Hash, Debug, Deserialize, Serialize)]
